pyproject.toml
[project]
name = "formsauthdecryptor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pycryptodome>=3.23.0",
]
